working with files and db in console and db tests were added

+++ src/main/java/mesnews/dao/AuteurDao.java
@@ -18,7 +18,7 @@ public class AuteurDao {
         return auteurs;
     }
 
-    public static Auteur readAuteur(Long id) {
+    public static Auteur readAuteur(int id) {
         SessionFactory sf = HibernateUtil.getSessionFactory();
         Session session = sf.openSession();
         Auteur auteur = (Auteur) session.get(Auteur.class, id);
